+++ src/Python/kittie/__init__.py
@@ -1,2 +1,3 @@
 #from kittie import *
 from kittie.kittie import *
+import kittie.kittie_common as kittie_common
