+++ src/Python/kittie/kittie_common.py
@@ -1,0 +1,1 @@
+../../../util/kittie_common.py
